Maplibre map works on iOS and Android
